Add files via upload
This is the code in C++ of the game 'Tower of Hanoi'. I have written this code in C++ using basic data structure (Stack), as I this project is the beginning of my coding journey, my code might contain bugs, errors and changes. Feel free for proving me your suggestions and opinions over my code. Feel free to enjoy the code and play the game by running the code. Thanks!!